Module changes to prep for gcc/make compatibility
